+++ CHANGELOG.md
@@ -10,10 +10,15 @@ context data may observe a performance drop when returning multiple objects. To 
 Config Context is not needed, the query parameter `?exclude=config_context` may be added to the request as to remove
 the Config Context from being included in any results.
 
+### Tag Permissions Changed
+
+NetBox now makes use of its own `Tag` model instead of the vanilla model which ships with django-taggit. This new model lives in the `extras` app and thus any permissions that you may have configured using "Taggit | Tag" should be changed to now use "Extras | Tag."
+
 ## Enhancements
 
 * [#2324](https://github.com/digitalocean/netbox/issues/2324) - Add color option for tags
 * [#2791](https://github.com/digitalocean/netbox/issues/2791) - Add a comment field for tags
+* [#2926](https://github.com/digitalocean/netbox/issues/2926) - Add changelog to the Tag model
 
 ---